undo the static casts as the return type changed again

+++ src/Cpp/1-getting-started/1-3-3-SettingUp3DRendering/Setting3DApplication.cpp
@@ -496,7 +496,7 @@ void Setting3DApplication::Render()
 
     D3D11_BUFFER_DESC description = {};
     _triangleVertices->GetDesc(&description);
-    UINT stride = static_cast<UINT>(_shaderCollection.GetLayoutByteSize(VertexType::PositionColorUv));
+    UINT stride = _shaderCollection.GetLayoutByteSize(VertexType::PositionColorUv);
     _deviceContext->IASetVertexBuffers(
         0,
         1,
